refactor: rename context to contexts and add path alias
