+++ src/components/forms/connector/source/CreateSourceForm/CreateSourceForm.tsx
@@ -13,7 +13,7 @@ import {
 import { useCreateSource, useSources } from "@/services/connector";
 import { CreateSourcePayload } from "@/lib/instill";
 import { Nullable } from "@/types/general";
-import { useAmplitudeCtx } from "context/AmplitudeContext";
+import { useAmplitudeCtx } from "@/contexts/AmplitudeContext";
 import { sendAmplitudeData } from "@/lib/amplitude";
 
 export type CreateSourceFormValues = {
