+++ src/components/forms/model/CreateModelForm/CreateModelForm.tsx
@@ -33,7 +33,7 @@ import {
   Model,
 } from "@/lib/instill";
 import { Nullable } from "@/types/general";
-import { useAmplitudeCtx } from "context/AmplitudeContext";
+import { useAmplitudeCtx } from "@/contexts/AmplitudeContext";
 import { sendAmplitudeData } from "@/lib/amplitude";
 import { AxiosError } from "axios";
 import { ErrorDetails } from "@/lib/instill/types";
