+++ src/components/forms/pipeline/ConfigurePipelineForm/ConfigurePipelineForm.tsx
@@ -10,7 +10,7 @@ import { PrimaryButton } from "@/components/ui";
 import { Pipeline, PipelineState } from "@/lib/instill";
 import { Nullable } from "@/types/general";
 import { useDeletePipeline, useUpdatePipeline } from "@/services/pipeline";
-import { useAmplitudeCtx } from "context/AmplitudeContext";
+import { useAmplitudeCtx } from "@/contexts/AmplitudeContext";
 import { sendAmplitudeData } from "@/lib/amplitude";
 import { DeleteResourceModal } from "@/components/modals";
 import { useRouter } from "next/router";
