+++ src/components/forms/pipeline/CreatePipelineForm/SetupDestinationStep/UseExistingDestinationFlow.tsx
@@ -9,7 +9,7 @@ import {
   CreatePipelineFormValues,
 } from "../CreatePipelineForm";
 import { useDestinations } from "@/services/connector";
-import { useAmplitudeCtx } from "context/AmplitudeContext";
+import { useAmplitudeCtx } from "@/contexts/AmplitudeContext";
 import { sendAmplitudeData } from "@/lib/amplitude";
 
 export type UseExistingDestinationFlowProps = StepNumberState;
